[Update] 2023 Core3 merger

+++ MMOCoreORB/src/server/ServerCore.cpp
@@ -35,6 +35,8 @@
 #include "engine/service/ServiceThread.h"
 #include "engine/lua/LuaPanicException.h"
 
+#include "server/zone/managers/statistics/StatisticsManager.h"
+
 ManagedReference<ZoneServer*> ServerCore::zoneServerRef = nullptr;
 SortedVector<String> ServerCore::arguments;
 bool ServerCore::truncateAllData = false;
@@ -641,6 +643,8 @@ void ServerCore::signalShutdown(ShutdownFlags flags) {
 }
 
 void ServerCore::initialize() {
+	StatisticsManager::instance()->markCoreStart(Thread::getProcessID());
+
 	info(true) << "Server start, pid: "
 		<< Thread::getProcessID() << ", time: " << Time().getFormattedTime();
 
@@ -679,7 +683,10 @@ void ServerCore::initialize() {
 		if (configManager->getMakeLogin()) {
 			loginServer = new LoginServer(configManager);
 			loginServer->deploy("LoginServer");
-			database->instance()->executeStatement("TRUNCATE TABLE sessions");
+
+			if (!ConfigManager::instance()->getLoginEnableSessionId()) {
+				database->instance()->executeStatement("TRUNCATE TABLE sessions");
+			}
 		}
 
 		if (configManager->getMakeZone()) {
@@ -774,6 +781,8 @@ void ServerCore::initialize() {
 		}
 
 		ObjectManager::instance()->scheduleUpdateToDatabase();
+
+		StatisticsManager::instance()->markCoreInitialized();
 
 		info("initialized", true);
